renaming "LoadingSpinner" and adjusting imports
