+++ frontend/src/metabase/common/components/DelayedLoading/index.ts
@@ -1,0 +1,1 @@
+export * from "./DelayedLoading";
