+++ frontend/src/metabase/common/components/Pickers/EntityPicker/components/index.tsx
@@ -1,5 +1,5 @@
 export * from "./AutoScrollBox";
 export * from "./EntityPickerModal";
 export * from "./ItemLists";
-export * from "./LoadingSpinner";
+export * from "./ItemListLoader";
 export * from "./NestedItemPicker";
